requirements.txt
Django==1.7.1
Pillow==2.6.1
argparse==1.2.1
bootstrap-admin==0.3.3
decorator==4.0.6
django-ckeditor-updated==4.4.4
gunicorn==19.3.0
ipython==4.1.1
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
ptyprocess==0.5.1
simplegeneric==0.8.1
sorl-thumbnail==11.12.1b
traitlets==4.1.0
wsgiref==0.1.2
